+++ cli/src/config.rs
@@ -24,8 +24,10 @@ use std::process::Command;
 use std::sync::Arc;
 use std::sync::LazyLock;
 use std::sync::Mutex;
+use std::time::SystemTime;
 
 use etcetera::BaseStrategy as _;
+use indoc::indoc;
 use itertools::Itertools as _;
 use jj_lib::config::ConfigFile;
 use jj_lib::config::ConfigGetError;
@@ -39,6 +41,9 @@ use jj_lib::config::ConfigValue;
 use jj_lib::config::StackedConfig;
 use jj_lib::dsl_util::AliasDeclarationParser;
 use jj_lib::dsl_util::AliasesMap;
+use jj_lib::file_util::IoResultExt as _;
+use jj_lib::protos::secure_config::ConfigMetadata;
+use jj_lib::protos::secure_config::TrustLevel;
 use jj_lib::secure_config::LoadedSecureConfig;
 use jj_lib::secure_config::SecureConfig;
 use rand::SeedableRng as _;
@@ -51,13 +56,16 @@ use tracing::instrument;
 use crate::command_error::CommandError;
 use crate::command_error::config_error;
 use crate::command_error::config_error_with_message;
+use crate::command_error::internal_error;
 use crate::ui::Ui;
+use crate::ui::UiStdout;
 
 // TODO(#879): Consider generating entire schema dynamically vs. static file.
 pub const CONFIG_SCHEMA: &str = include_str!("config-schema.json");
 
 const REPO_CONFIG_DIR: &str = "repos";
 const WORKSPACE_CONFIG_DIR: &str = "workspaces";
+pub const MANAGED_CONFIG_PATH: &str = ".config/jj/config.toml";
 
 /// Parses a TOML value expression. Interprets the given value as string if it
 /// can't be parsed and doesn't look like a TOML expression.
@@ -598,6 +606,149 @@ impl ConfigEnv {
         {
             config.as_mut().load_file(ConfigSource::Workspace, path)?;
         }
+        Ok(())
+    }
+
+    #[instrument(skip(ui))]
+    pub fn reload_managed_config(
+        &self,
+        ui: &Ui,
+        config: &mut RawConfig,
+    ) -> Result<(), CommandError> {
+        config.as_mut().remove_layers(ConfigSource::Managed);
+        let config_file = if let Some(path) = self.workspace_path.as_deref() {
+            path.join(MANAGED_CONFIG_PATH)
+        } else {
+            return Ok(());
+        };
+        if !config_file.try_exists()? {
+            return Ok(());
+        }
+        let (repo_config_path, mut metadata) = if let Some(loaded_config) =
+            self.load_secure_config(ui, self.repo_config.as_ref(), REPO_CONFIG_DIR, true)?
+        {
+            // Cannot be None since we forced it to create the repo config.
+            (loaded_config.config_file.unwrap(), loaded_config.metadata)
+        } else {
+            return Ok(());
+        };
+        match TrustLevel::try_from(metadata.trust_level) {
+            Ok(TrustLevel::Ignored) => {}
+            Ok(TrustLevel::Trusted) => config
+                .as_mut()
+                .load_file(ConfigSource::Managed, &config_file)?,
+            Ok(TrustLevel::Notify) => {
+                // We already checked config_file.try_exists(), so it's
+                // reasonable to error out here if it fails.
+                let working_copy_mtime = std::fs::metadata(&config_file)
+                    .context(&config_file)?
+                    .modified()
+                    .context(&config_file)?;
+                let repo_config_mtime = match self.repo_config_path(ui)?.map(|path| {
+                    std::fs::metadata(&path)
+                        .context(&path)?
+                        .modified()
+                        .context(&path)
+                }) {
+                    Some(Ok(mtime)) => mtime,
+                    Some(Err(e)) if e.source.kind() != std::io::ErrorKind::NotFound => {
+                        return Err(e)?;
+                    }
+                    // If the file wasn't found, treat it as if it's really old and thus out of
+                    // date.
+                    _ => SystemTime::UNIX_EPOCH,
+                };
+                if repo_config_mtime < working_copy_mtime {
+                    writeln!(
+                        ui.warning_default(),
+                        indoc! {"
+                            The managed config has been updated more recently than the repo config.
+                            Please review the changes in {} and copy any desired settings to your repo config (`jj config edit --repo`).
+                        "},
+                        config_file.display(),
+                    )?;
+                }
+            }
+            Ok(TrustLevel::Unset) | Err(_) => {
+                if Ui::can_prompt() && !matches!(ui.stdout(), UiStdout::Null(_)) {
+                    metadata.trust_level = match ui.prompt_choice(
+                        &format!(
+                            indoc! {"
+                                The repo at {} contains a managed config file. You can:
+                                (i) Ignore managed configs for this repo
+                                (n) Be notified when the managed config is more recent than your repo config (you keep your repo config up to date manually).
+                                (t) Trust this repo (any future changes to the managed config will be automatically applied)
+                                Please select an option (i/n/t)"
+                            },
+                            self.repo_path.as_ref().unwrap().display(),
+                        ),
+                        &["i", "n", "t"],
+                        None,
+                    )? {
+                        0 => TrustLevel::Ignored,
+                        1 => TrustLevel::Notify,
+                        2 => TrustLevel::Trusted,
+                        _ => unreachable!(),
+                    }
+                    .into();
+                    self.repo_config
+                        .as_ref()
+                        .unwrap()
+                        .update_metadata(&repo_config_path, metadata)?;
+                    // The config has now changed on disk, so just retry.
+                    return self.reload_managed_config(ui, config);
+                } else {
+                    // If we're unable to prompt the user, we should output a warning,
+                    // but for security reasons, we should assume ignore.
+                    writeln!(
+                        ui.warning_default(),
+                        indoc! {"
+                            The repo at {} contains a managed config file, but the managed config for this repo is not configured.
+                            Please run `jj config managed --ignore/notify/trust` to configure it
+                            * --ignore ignores managed configs for this repo
+                            * --notify notifies you when the managed config is more recent than your repo config (you keep your repo config up to date with the managed config manually).
+                            * --trust trusts this repo (any future changes to the managed config will be automatically applied)"
+                        },
+                        self.repo_path.as_ref().unwrap().display(),
+                    )?;
+                }
+            }
+        }
+        Ok(())
+    }
+
+    pub fn managed_config_path(&self) -> Option<PathBuf> {
+        self.workspace_path
+            .as_deref()
+            .map(|p| p.join(MANAGED_CONFIG_PATH))
+    }
+
+    pub fn managed_config_files(
+        &self,
+        config: &RawConfig,
+    ) -> Result<Vec<ConfigFile>, CommandError> {
+        config_files_for(config, ConfigSource::Managed, || {
+            Ok(self
+                .managed_config_path()
+                .map(|p| ConfigFile::load_or_empty(ConfigSource::Managed, p))
+                .transpose()?)
+        })
+    }
+
+    pub fn update_repo_metadata(
+        &self,
+        ui: &Ui,
+        update: impl FnOnce(&mut ConfigMetadata),
+    ) -> Result<(), CommandError> {
+        let mut loaded = self
+            .load_secure_config(ui, self.repo_config.as_ref(), REPO_CONFIG_DIR, true)?
+            .ok_or_else(|| internal_error("Failed to load repo config"))?;
+        update(&mut loaded.metadata);
+        // Cannot be None since we forced it to create the repo config.
+        self.repo_config
+            .as_ref()
+            .unwrap()
+            .update_metadata(&loaded.config_file.unwrap(), loaded.metadata)?;
         Ok(())
     }